Added unit and e2e tests
