+++ tests/unit/test_update_movie.py
@@ -1,6 +1,6 @@
 # TODO: Feature 5
-from app import app
 from src.repositories.movie_repository import get_movie_repository
+import pytest
 
 movie_repository = get_movie_repository()
 
@@ -11,34 +11,10 @@ def test_update_movie():
     assert (movie_repository.get_movie_by_id(original_movie.movie_id)).title == "New title"
     assert (movie_repository.get_movie_by_id(original_movie.movie_id)).director == "New dir"
     assert (movie_repository.get_movie_by_id(original_movie.movie_id)).rating == 3
+    assert original_movie.movie_id==updated_movie.movie_id
 
-def test_update_movie_flask():
-    test_app = app.test_client()
-    movie_repository.clear_db()
-    original_movie = movie_repository.create_movie("Test Title", "Test Director", 5)
-    id = original_movie.movie_id
-    response = test_app.post(f'/movies/{id}', data={
-        "title": "New title",
-        "director": "New dir",
-        "new_rating": 4
-    }, follow_redirects=True)
-    
-    updated_movie = movie_repository.get_movie_by_id(id)
-
-    assert updated_movie.title=="New title"
-    assert updated_movie.director=="New dir"
-    assert updated_movie.rating==4
-    assert response.status_code==200
-    
-def test_update_movie_bad():
-    test_app = app.test_client()
-    movie_repository.clear_db()
-    original_movie = movie_repository.create_movie("Test Title", "Test Director", 5)
-    bad_id = original_movie.movie_id + 1
-    response = test_app.post(f'/movies/{bad_id}', data={
-        "title": "New title",
-        "director": "New dir",
-        "new_rating": 4
-    })
-
-    assert response.status_code==500
+def test_update_movie_DNE():
+    with pytest.raises(ValueError) as info:
+        movie_repository.clear_db()
+        no_movie = movie_repository.update_movie(12345, "imaginary movie", "imaginary friend", 3)
+    assert "not found" in str(info.value)
